refactor(command.go; websocket_stream.go): make new filex
