+++ src/.env.template
@@ -1,1 +1,1 @@
-BOT_TOKEN =
+BOT_TOKEN =
